feat(interface-type-control.tsx): Add icons to interface type selection
