--- src/pages/main-page/components/http-display/components/interface-type-control/interface-type-control.tsx
@@ -1,29 +1,0 @@
-import { Select, MenuItem } from '@mui/material';
-import { Control, Controller } from 'react-hook-form';
-import { InterfaceInputs, InterfaceType } from '../../../../types';
-
-type InterfaceTypeControlProps<T extends InterfaceInputs> = {
-  control: Control<T>;
-};
-
-const InterfaceTypeControl = <T extends InterfaceInputs>({
-  control,
-}: InterfaceTypeControlProps<T>) => {
-  return (
-    <Controller
-      name="interfaceType"
-      control={control as unknown as Control<InterfaceInputs>}
-      render={({ field }) => (
-        <Select required {...field}>
-          {Object.keys(InterfaceType).map((interfaceType) => (
-            <MenuItem key={interfaceType} value={interfaceType}>
-              {InterfaceType[interfaceType as keyof typeof InterfaceType]}
-            </MenuItem>
-          ))}
-        </Select>
-      )}
-    />
-  );
-};
-
-export default InterfaceTypeControl;
